Don't load jQuery through require at all
This seems bizarre to me, but it’s actually what the flight docs (seem
to) suggest. Flight doesn’t declare jQuery as a dependency, even though
it is, so no way to alert requirejs of that (except maybe making an
intermediate module). Simplest solution is to just load jQuery in the
html before flight, as I’m doing now. Annoying, but not actually a big
perf hit, if any.