Bug fixes.
Attach leftSidebar to body so we don’t have to worry about whether
we’ve moved a component out of the header or not.

Delete old empty skrollr files

Start adding some console.log statements (they can be stripped in
production by uglify if desired)

DetermineAnimationMode now takes topMargin - nav into account, as it
should

Animations added for large-design pagination are removed for large
scrolling (this was a “duh”), and margin top added for large-design
scrolling is removed for large design pagination and the small design
scrolling (also a “duh”).

General refactoring