Use minified JS in production, but still allow raw js in dev
Commit 0d7088d944e96bef5384ac5bdfd0601fd1232d9d changed the template to
point to the minified JS. That was good for production, but it meant we
couldn’t run the faster `gulp dev` task during development (as that
task doesn’t create a minified js file, to save time + add legibility).
Now we can run `gulp dev` in development and the template will
reference the unminified scripts. But when the scripts are minified by
the normal gulp command, the template will also be modified dynamically
to link to the new, minified script.